Bump dep on @primer/react
The version we we re using takes a peer dep on react 17

This doesn't confuse yarn, but npm attempts to satisfy the deps
by installing react 17 instead of 18, and then shit fails to import.

Trying this out in order to see if the newest version solves the issue
because it takes a peer dep of 17 | 18.